Moved theme code into its own file, update imports
