+++ src/index.jsx
@@ -38,14 +38,14 @@ import {
   action as portfolioActivityViewerAction,
   PortfolioActivityViewer,
 } from "./Tools/_framework/Paths/PortfolioActivityViewer";
-import { ChakraProvider, extendTheme } from "@chakra-ui/react";
+import { ChakraProvider } from "@chakra-ui/react";
+import theme from "/home/node/workspace/src/theme.jsx";
 import {
   action as editorSupportPanelAction,
   loader as editorSupportPanelLoader,
 } from "./Tools/_framework/Panels/NewSupportPanel";
 import ErrorPage from "./Tools/_framework/Paths/ErrorPage";
 
-import "@fontsource/jost";
 import {
   PortfolioActivityEditor,
   loader as portfolioEditorLoader,
@@ -57,92 +57,6 @@ import {
   action as publicEditorAction,
 } from "./Tools/_framework/Paths/PublicEditor";
 // import { loader as portfolioEditorMenuCapLoader } from './Tools/_framework/MenuPanelCaps/PortfolioEditorInfoCap';
-
-const theme = extendTheme({
-  fonts: {
-    body: "Jost",
-  },
-  textStyles: {
-    primary: {
-      fontFamily: "Jost",
-    },
-  },
-  config: {
-    initialColorMode: "light",
-    useSystemColorMode: false,
-    // initialColorMode: "system",
-    // useSystemColorMode: true,
-  },
-  colors: {
-    doenet: {
-      mainBlue: "#1a5a99",
-      lightBlue: "#b8d2ea",
-      solidLightBlue: "#8fb8de",
-      mainGray: "#e3e3e3",
-      mediumGray: "#949494",
-      lightGray: "#e7e7e7",
-      donutBody: "#eea177",
-      donutTopping: "#6d4445",
-      mainRed: "#c1292e",
-      lightRed: "#eab8b8",
-      mainGreen: "#459152",
-      canvas: "#ffffff",
-      canvastext: "#000000",
-      lightGreen: "#a6f19f",
-      lightYellow: "#f5ed85",
-      whiteBlankLink: "#6d4445",
-      mainYellow: "#94610a",
-      mainPurple: "#4a03d9",
-    },
-  },
-  components: {
-    Button: {
-      baseStyle: {
-        fontWeight: "normal",
-        letterSpacing: ".5px",
-        // _focus: {
-        //   outline: "2px solid #2D5994",
-        //   outlineOffset: "2px",
-        // },
-        // _disabled: {
-        //   bg: "#E2E2E2",
-        //   color: "black",
-        // },
-      },
-      variants: {
-        // We can override existing variants
-        solid: {
-          bg: "doenet.mainBlue",
-          color: "white",
-          _hover: {
-            bg: "doenet.solidLightBlue",
-            color: "black",
-            _disabled: {
-              bg: "doenet.mainBlue",
-              color: "white",
-            },
-          },
-        },
-        outline: {
-          borderColor: "doenet.mainBlue",
-          _hover: {
-            bg: "doenet.solidLightBlue",
-          },
-        },
-        ghost: {
-          _hover: {
-            bg: "doenet.solidLightBlue",
-          },
-        },
-        link: {
-          color: "doenet.mainBlue",
-        },
-      },
-    },
-  },
-});
-
-export default theme;
 
 const router = createBrowserRouter([
   {
